Refactor `DevLakeUserAndTeamsConfig` to `UserAndTeamsConfig`, update references, and simplify related code.
