+++ utilities/src/jvmMain/kotlin/com/github/karlsabo/devlake/UserAndTeamsConfig.kt
@@ -2,14 +2,12 @@ package com.github.karlsabo.devlake
 
 import com.github.karlsabo.devlake.accessor.Team
 import com.github.karlsabo.devlake.accessor.TeamUser
-import com.github.karlsabo.devlake.accessor.User
-import com.github.karlsabo.devlake.accessor.UserAccount
+import com.github.karlsabo.dto.User
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class DevLakeUserAndTeamsConfig(
+data class UserAndTeamsConfig(
     val users: List<User>,
-    val userAccounts: List<UserAccount>,
     val teams: List<Team>,
-    val teamUsers: List<TeamUser>
+    val teamUsers: List<TeamUser>,
 )
